requirements.txt
asgiref==3.8.1
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
